# Optimize extract_clusters and prepare_clusters systems