@rafalh 2023-11-27
> Can you run this on some of our more intensive stress test examples? In this particular case, `many_lights` might be a better test than `3d_scene`.

Here you are - tracy comparison for `many_lights`:
![image](https://github.com/bevyengine/bevy/assets/160391/d4854111-231d-430f-b84e-5f1b2e17b0c1)

I showed data for `3d_scene` first (which may be not something that we want to optimize) because it has small amount of lights (just 1) that has entire scene in range. As I understand it it is a pessimistic case where nearly all clusters are in range of a point light causing a lot of `Vec` allocations.
In `many_lights` example we have a lot of very small lights (radius = 0.3) so most of the clusters are empty and in such case cloning vectors is not a bottleneck.
I am not sure how much this optimization would affect real projects. I think it may affect some with indoor areas or if someone tries to model sun as a point light. But even if it was for simple examples like `3d_scene` I think it is worth merging because of the impact on FPS. Low FPS (when compared to other engines) in simple scenes may discourage people from using Bevy.